+++ neo/tools/common/PropertyGrid.cpp
@@ -75,7 +75,7 @@ bool rvPropertyGrid::Create ( HWND parent, int id, int style )
 	mStyle = style;
 
 	// Create the List view
-	mWindow = CreateWindowEx ( 0, "LISTBOX", "", WS_VSCROLL|WS_CHILD|WS_VISIBLE|LBS_OWNERDRAWFIXED|LBS_NOINTEGRALHEIGHT|LBS_NOTIFY, 0, 0, 0, 0, parent, (HMENU)id, win32.hInstance, 0 );	
+	mWindow = CreateWindowEx ( 0, "LISTBOX", "", WS_VSCROLL|WS_CHILD|WS_VISIBLE|LBS_OWNERDRAWFIXED|LBS_NOINTEGRALHEIGHT|LBS_NOTIFY, 0, 0, 0, 0, parent, (HMENU)(intptr_t)id, win32.hInstance, 0 );	
 	mListWndProc = (WNDPROC)GetWindowLongPtr ( mWindow, GWLP_WNDPROC );
 	SetWindowLongPtr ( mWindow, GWLP_USERDATA, (LONG_PTR)this );
 	SetWindowLongPtr ( mWindow, GWLP_WNDPROC, (LONG_PTR)WndProc );
