+++ neo/tools/compilers/roqvq/roqParam.cpp
@@ -157,7 +157,7 @@ void roqParam::InitFromFile( const char *fileName )
 		if (token.Icmp( "filename") == 0) {
 			src->ReadToken( &token );
 			outputFilename = token;
-			i = strlen(outputFilename);
+			i = (int)strlen(outputFilename);
 //			common->Printf("  + output file is %s\n", outputFilename );
 			readarg++;
 			continue;
@@ -578,7 +578,7 @@ int parseRange(const char *rangeStr,int field, int skipnum[], int startnum[], in
 	realnum += numfiles[field];
 	if ( start[0] == '0' && start[1] != '\0' ) {
 		padding[field] = true;
-		numpadding[field] = strlen( start );
+		numpadding[field] = (int)strlen( start );
 	} else {
 		padding[field] = false;
 	}
@@ -627,7 +627,7 @@ int parseTimecodeRange(const char *rangeStr,int field, int skipnum[], int startn
 	realnum += numfiles[field];
 	if ( start[0] == '0' && start[1] != '\0' ) {
 		padding[field] = true;
-		numpadding[field] = strlen( start );
+		numpadding[field] = (int)strlen( start );
 	} else {
 		padding[field] = false;
 	}
